+++ source/middle-spec.ts
@@ -1,16 +1,16 @@
-import { map, pipe, init } from 'rambda'
+import { map, middle, pipe } from 'rambda'
 
-describe('R.init', () => {
+describe('R.middle', () => {
   it('with string', () => {
-    const result = init('foo')
+    const result = middle('foo')
 
     result // $ExpectType string
   })
   it('with list - using const on short array', () => {
     const result = pipe(
-      [1] as const,
+      [1, 2] as const,
       map(x => x * 2),
-      init,
+      middle,
     )
     result // $ExpectType []
   })
@@ -18,20 +18,20 @@ describe('R.init', () => {
     const result = pipe(
       [] as const,
       map(x => x * 2),
-      init,
+      middle,
     )
     result // $ExpectType number[]
   })
   it('with list - using const', () => {
     const result = pipe(
-      [1, 2, 3] as const,
+      [1, 2, 3, 4] as const,
       map(x => x * 2),
-      init,
+      middle,
     )
     result // $ExpectType [number, number]
   })
   it('with list - mixed types', () => {
-    const result = init(['foo', 'bar', 1, 2, 3])
+    const result = middle(['foo', 'bar', 1, 2, 3])
 
     result // $ExpectType (string | number)[]
   })