+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 - Add `R.filterMap` - similar to Ruby `filter_map`
 
+- Add `R.mapChain` - when in `R.pipe` there are several `R.map` one after the other, then `R.mapChain` can be used instead.
+
+- Add `R.middle` - equal to `R.init` + `R.tail`
+
+- Add `R.sumByProp`
+
+- Add `R.random`, `R.shuffle`, `R.switcher`, `R.sum`, `R.delay` - imported from `Rambda`
+
 - Add index to `R.filter`/`R.reject` predicate signiture
+
+- Improve typing of `R.init`, `R.tail`
 
 11.0.1